historic/tankensub/requirements.txt
paho-mqtt==2.0.0
proximityhash
python-geohash
cdbx
